--- routes/contact.js
@@ -1,7 +1,0 @@
-const express = require('express');
-const router = express.Router();
-const { senderEmailContact } = require('../controllers/contactController');
-
-router.post('/', senderEmailContact);
-
-module.exports = router;
