+++ routes/contacts.js
@@ -4,11 +4,12 @@ const {
   addContact,
   getAllContacts,
 } = require('../controllers/contactController');
+const { senderEmailContact } = require('../service/EmailService');
 const validatingContactValues = require('../middlewares/contactValidation');
 const { isAdmin } = require('../middlewares/verify');
 
 router.post('/', validatingContactValues, addContact);
-
 router.get('/', isAdmin, getAllContacts);
+router.post('/', senderEmailContact);
 
 module.exports = router;
